Rename task-badge-variants
